chore: convert mongodb test into testable example
